requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.2
Django==5.1.6
django-allauth==65.4.1
django-appconf==1.1.0
django-browser-reload==1.18.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-jazzmin==3.0.1
django-tailwind==3.5.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
